updates to 09_sequence/sinewaves and 09/sequence/temperature plus labs
