--- courses/machine_learning/deepdive/09_sequence/sinemodel/PKG-INFO
@@ -1,10 +1,0 @@
-Metadata-Version: 1.0
-Name: trainer
-Version: 0.1
-Summary: Google Cloud Machine Learning Example
-Home-page: UNKNOWN
-Author: Google
-Author-email: lak@google.com
-License: UNKNOWN
-Description: UNKNOWN
-Platform: UNKNOWN
