--- courses/machine_learning/deepdive/09_sequence/sinemodel/setup.cfg
@@ -1,5 +1,0 @@
-[egg_info]
-tag_build = 
-tag_date = 0
-tag_svn_revision = 0
-
