--- courses/machine_learning/deepdive/09_sequence/sinemodel/trainer/.gitignore
@@ -1,1 +1,0 @@
-*.pyc
